fix(platform): fix connector page interaction tests
- move test file from views/conn/__tests__ to views/zero-page/__tests__ to co-locate with existing connector tests
- remove CONN-I-010 (duplicate of existing search filter test)
- remove CONN-I-014 (duplicate subset of existing disconnect dropdown tests)
- remove CONN-I-015 (duplicate of existing disconnect scenario)
- fix CONN-I-011: replace ui text assertion with behavioral assertion (Save button in api-token form)
- fix CONN-I-013: replace modal heading assertion with behavioral assertion (scope items rendered)
- fix CONN-I-013 msw response to include all required schema fields
