+++ turbo/apps/platform/src/views/zero-page/__tests__/zero-connectors-page-interaction.test.tsx
@@ -1,0 +1,94 @@
+/**
+ * Interaction tests for the /connectors page (ZeroConnectorsPage component).
+ *
+ * Tests user interactions (connect, reconnect, review) via setupPage following
+ * platform testing principles:
+ * - Entry point: setupPage({ path: "/connectors" })
+ * - Mock (external): Web API via MSW
+ * - Real (internal): All signals, components, rendering
+ */
+
+import { afterEach, expect, test, vi } from "vitest";
+import { screen, waitFor } from "@testing-library/react";
+import userEvent from "@testing-library/user-event";
+import { http, HttpResponse } from "msw";
+import { server } from "../../../mocks/server.ts";
+import { testContext } from "../../../signals/__tests__/test-helpers.ts";
+import { setupPage } from "../../../__tests__/page-helper.ts";
+import { mockConnectors } from "./zero-connectors-page-test-helpers.ts";
+
+const context = testContext();
+
+afterEach(() => {
+  vi.restoreAllMocks();
+});
+
+test("connect button opens api-token form (CONN-I-011)", async () => {
+  const user = userEvent.setup();
+  await setupPage({ context, path: "/connectors" });
+
+  await waitFor(() => {
+    expect(
+      screen.getByRole("button", { name: "Connect Axiom" }),
+    ).toBeInTheDocument();
+  });
+
+  await user.click(screen.getByRole("button", { name: "Connect Axiom" }));
+
+  await waitFor(() => {
+    expect(screen.getByRole("dialog")).toBeInTheDocument();
+  });
+  expect(screen.getByRole("button", { name: "Save" })).toBeInTheDocument();
+});
+
+test("reconnect button triggers reconnection flow (CONN-I-012)", async () => {
+  const user = userEvent.setup();
+  mockConnectors([{ type: "github", needsReconnect: true }]);
+
+  const openSpy = vi
+    .spyOn(window, "open")
+    .mockReturnValue({ closed: false } as Window);
+
+  await setupPage({ context, path: "/connectors" });
+
+  await waitFor(() => {
+    expect(screen.getByText("Reconnect")).toBeInTheDocument();
+  });
+
+  await user.click(screen.getByText("Reconnect"));
+
+  expect(openSpy).toHaveBeenCalledWith(
+    expect.stringContaining("/api/zero/connectors/github/authorize"),
+    "_blank",
+    expect.any(String),
+  );
+});
+
+test("review button opens scope diff with added permissions (CONN-I-013)", async () => {
+  const user = userEvent.setup();
+  mockConnectors([{ type: "github", oauthScopes: [] }]);
+
+  server.use(
+    http.get("*/api/zero/connectors/:type/scope-diff", () => {
+      return HttpResponse.json({
+        addedScopes: ["repo", "project"],
+        removedScopes: [],
+        currentScopes: [],
+        storedScopes: ["repo", "project"],
+      });
+    }),
+  );
+
+  await setupPage({ context, path: "/connectors" });
+
+  await waitFor(() => {
+    expect(screen.getByText("Review")).toBeInTheDocument();
+  });
+
+  await user.click(screen.getByText("Review"));
+
+  await waitFor(() => {
+    expect(screen.getByText("repo")).toBeInTheDocument();
+  });
+  expect(screen.getByText("project")).toBeInTheDocument();
+});
